make plugin handle registration of new Result classes on its own
This avoids needing a patch like the one for Koha bug 26163 in order to
run the plugin.

This does depend on Module::Metadata being present; since that's a core
Perl module as of 5.14, that should be a reasonably safe assumption except
on truly ancient systems that are somehow managing to run a recent version
of Koha.

Note that a koha-plack --restart will still be needed to get the new API
routes regisered.

Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>